Moving autosavespinner to a global  component
