+++ generators/app/templates/src/components/AutoSaveSpinner/auto-save.style.js
