+++ generators/app/templates/src/components/AutoSaveSpinner/index.js
