+++ generators/app/templates/src/components/index.js
@@ -5,5 +5,15 @@ import AuthNavBar from './AuthNavBar';
 import ImageProfile from './ImageProfile';
 import Footer from './Footer';
 import Notification from './Notifications';
+import AutoSaveSpinner from './AutoSaveSpinner';
 
-export { ErrorBoundary, GlobalError, NavBar, AuthNavBar, ImageProfile, Footer, Notification };
+export {
+  ErrorBoundary,
+  GlobalError,
+  NavBar,
+  AuthNavBar,
+  ImageProfile,
+  Footer,
+  Notification,
+  AutoSaveSpinner
+};
