--- generators/app/templates/src/containers/FormModel/Renderer/children/auto-save.style.js
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const AutoSaveSpinnerWrapper = styled.span`
-  margin-top: 20px;
-`;
